+++ synapse/models/gov.py
@@ -2,7 +2,7 @@
 icpregex = '^(皖|京|渝|闽|粤|甘|桂|黔|豫|鄂|冀|琼|港|黑|湘|吉|苏|赣|辽|澳|蒙|宁|青|川|鲁|沪|陕|晋|津|台|新|藏|滇|浙)ICP(备|证)[0-9]{8}号$'
 
 modeldefs = (
-    ('gov:cn', {
+    {
         'types': (
 
             ('gov:cn:icp', ('str', {'regex': icpregex}), {
@@ -21,8 +21,8 @@ modeldefs = (
             ('gov:cn:icp', {}, ()),
             ('gov:cn:mucd', {}, ()),
         )
-    }),
-    ('gov:intl', {
+    },
+    {
         'types': (
             ('iso:oid', ('str', {'regex': '^([0-2])((\\.0)|(\\.[1-9][0-9]*))*$'}), {
                 'doc': 'An ISO Object Identifier string.'}),
@@ -59,8 +59,8 @@ modeldefs = (
             ('iso:3166:alpha3', {}, ()),
             ('iso:3166:numeric3', {}, ()),
         ),
-    }),
-    ('gov:us', {
+    },
+    {
         'types': (
 
             ('gov:us:ssn', ('str', {'regex': '^[0-9]{3}-[0-9]{2}-[0-9]{4}'}), {
@@ -118,5 +118,5 @@ modeldefs = (
             ('gov:us:ssn', {}, []),
             ('gov:us:zip', {}, []),
         ),
-    }),
+    },
 )